+++ lib/sgdisk.rb
@@ -42,13 +42,13 @@ module Sgdisk
     def make_boot
       @boot || return
 
-      partition @boot, @boot_code, '0:+256MiB'
+      partition @boot, @boot_code, "0:+#{Getch::OPTIONS[:boot_size]}MiB"
     end
 
     def make_swap
       @swap || return
 
-      mem = Getch::Helpers.get_memory
+      mem = "#{Getch::OPTIONS[:swap_size]}M"
       partition @swap, @swap_code, "0:+#{mem}"
     end
 
@@ -98,7 +98,7 @@ module Sgdisk
       disk = dev[/^[a-z]+/]
       cmd = Getch::Command.new("sgdisk -E /dev/#{disk}")
       end_position = cmd.res.to_i
-      ( end_position - ( end_position + 1 ) % 2048 )
+      (end_position - (end_position + 1) % 2048)
     end
   end
 
@@ -135,11 +135,11 @@ module Sgdisk
     def make_boot
       @boot || return
 
-      partition @boot, @boot_code, '0:+2G'
+      partition @boot, @boot_code, "0:+#{Getch::OPTIONS[:boot_size]}MiB"
     end
 
     def make_swap
-      mem = Getch::Helpers.get_memory
+      mem = "#{Getch::OPTIONS[:swap_size]}M"
       partition @swap, @swap_code, "0:+#{mem}"
       add_zlog
       add_zcache